% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3.R
\name{init}
\alias{init}
\alias{primer3_init}
\alias{primer3_free}
\title{Loading Primer3 configuration files}
\usage{
primer3_init(config_path = system.file("extdata/primer3_config", package
  = "primer3"))

primer3_free()
}
\arguments{
\item{config_path}{Path to Primer3 configuration files. These are installed by default with
this package. The path must not end with a "\code{/}".}
}
\description{
For all functions besides \code{calculate_tm}, Primer3 requires initialization to
load the thermodynamic configuration data. The initialization will happen automatically
(and only once for each session) when a function that needs initialization is called.
(Or, the user can initialize manually by calling \code{primer3_init}, although there is
no advantage to this.) To avoid memory leaks, users should call \code{primer3_free} at
the end of the program to free memory allocated during initialization. Calling 
\code{primer3_free} after every call to a function will slow performance, since the 
configuration files will need to be reloaded during subsequent calls.
}
