% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3.R
\name{thermo}
\alias{thermo}
\alias{calculate_tm}
\alias{calculate_hairpin}
\alias{calculate_homodimer}
\alias{calculate_dimer}
\title{Thermodynamic calculations with Primer3}
\usage{
calculate_tm(oligos, salt_conc = 50, divalent_conc = 0,
  dntp_conc = 0, dna_conc = 50, nn_max_len = 60,
  tm_method = "SantaLucia", salt_correction = "SantaLucia")

calculate_hairpin(oligo, ...)

calculate_homodimer(oligo, ...)

calculate_dimer(oligo1, oligo2, ...)
}
\arguments{
\item{oligos}{Character vector of oligos. Calling once with multiple oligos is faster 
than repeated calls with single oligos.}

\item{salt_conc}{Monovalent salt concentration in mM.}

\item{divalent_conc}{Divalent ion concentration in mM.}

\item{dntp_conc}{dNTP concentration in mM.}

\item{dna_conc}{DNA concentration in nM.}

\item{nn_max_len}{Tm for oligos up to this length will be calculated using a nearest-neighbor
method. Beyond this length, the Tm is extrapolated based on GC content.}

\item{tm_method}{Method for Tm calculations. Options include "Breslauer" and "SantaLucia". 
As with \code{salt_correction}, this package uses the recommended Primer3
method ("SantaLucia"), not the Primer3 default method ("Schildkraut").}

\item{salt_correction}{Method for salt correction calculations. Options include "Schildkraut", 
"SantaLucia", "Owczarzy". As with \code{tm_method}, this package uses the recommended Primer3
method ("SantaLucia"), not the Primer3 default method ("Schildkraut").}

\item{maxloop}{Maximum length of loop structures. Not available for \code{calculate_tm}.}

\item{temp_c}{Temperature in degrees Celsius. Not available for \code{calculate_tm}.}

\item{print_output}{If \code{TRUE}, print alignment or secondary structure to terminal. 
When \code{TRUE}, no output is returned. Not available for \code{calculate_tm}.}
}
\value{
For \code{calculate_tm}, a numeric vector with the melting temperature [C] for each
input oligo. For the other functions, a named list of vectors indicating if a structure was 
found (\code{structure_found}); the changes in entropy (\code{ds}), enthalpy (\code{dh}), 
and Gibbs free energy (\code{dg}); and the alignment locations for each end (\code{align_end_1}
and \code{align_end_2}). Note that if \code{print_output} is \code{TRUE}, the functions return
\code{NULL}.
}
\description{
These functions allow direct access to the Primer3 thermodynamic libraries via Rcpp.
For all functions besides \code{calculate_tm}, Primer3 requires initialization to
load the thermodynamic configuration data. The initialization will happen automatically
(and only once for each session) when a function that needs initialization is called.
(Or, the user can initialize manually by calling \code{primer3_init}, although there is
no advantage to this.) To avoid memory leaks, users should call \code{primer3_free} at
the end of the program to free memory allocated during initialization. Calling 
\code{primer3_free} after every call to a function will slow performance, since the 
configuration files will need to be reloaded during subsequent calls.
}
\seealso{
\code{\link{primer3_free}}
}
